Add fetch-images.sh, cdi manifests
Create a script to fetch all images required from either the provisioning scripts or the
manifests in the subfolder. Will be used to populate an image list file for pre-pulling
fixed version checking and after cluster-up checking actual deployed images.

Also add cdi manifests as they are a test dependency for kubevirt.

Signed-off-by: Daniel Hiller <daniel.hiller.1972@gmail.com>